Refactor plasma exhaust calculations to plasma fuelling, updating related plotting functions for tritium, deuterium, and alpha particle flow rates
